+++ examples/client_messenger.py
@@ -1,0 +1,86 @@
+"""
+Messenger example client
+
+Bridges minecraft chat (in/out) with stdout and stdin.
+"""
+
+import os
+import sys
+
+from twisted.internet import defer, reactor, stdio
+from twisted.protocols import basic
+from quarry.auth import ProfileCLI
+from quarry.net.client import ClientFactory, SpawningClientProtocol
+
+
+class StdioProtocol(basic.LineReceiver):
+    delimiter = os.linesep.encode('ascii')
+    in_encoding  = getattr(sys.stdin,  "encoding", None)
+    out_encoding = getattr(sys.stdout, "encoding", None)
+
+    def lineReceived(self, line):
+        self.minecraft_protocol.send_chat(line.decode(self.in_encoding))
+
+    def send_line(self, text):
+        self.sendLine(text.encode(self.out_encoding))
+
+
+class MinecraftProtocol(SpawningClientProtocol):
+    spawned = False
+
+    def packet_chat_message(self, buff):
+        p_text = buff.unpack_chat()
+
+        # 1.7.x
+        if self.protocol_version <= 5:
+            p_position = 0
+        # 1.8.x
+        else:
+            p_position = buff.unpack('B')
+
+        if p_position in (0, 1) and p_text.strip():
+            self.stdio_protocol.send_line(p_text)
+
+    def send_chat(self, text):
+        self.send_packet("chat_message", self.buff_type.pack_string(text))
+
+
+class MinecraftFactory(ClientFactory):
+    protocol = MinecraftProtocol
+    log_level = "WARN"
+
+    def buildProtocol(self, addr):
+        minecraft_protocol = super(MinecraftFactory, self).buildProtocol(addr)
+        stdio_protocol = StdioProtocol()
+
+        minecraft_protocol.stdio_protocol = stdio_protocol
+        stdio_protocol.minecraft_protocol = minecraft_protocol
+
+        stdio.StandardIO(stdio_protocol)
+        return minecraft_protocol
+
+
+@defer.inlineCallbacks
+def run(args):
+    # Log in
+    profile = yield ProfileCLI.make_profile(args)
+
+    # Create factory
+    factory = MinecraftFactory(profile)
+
+    # Connect!
+    yield factory.connect(args.host, args.port)
+
+
+def main(argv):
+    parser = ProfileCLI.make_parser()
+    parser.add_argument("host")
+    parser.add_argument("port", nargs='?', default=25565, type=int)
+    args = parser.parse_args(argv)
+
+    run(args)
+    reactor.run()
+
+
+if __name__ == "__main__":
+    main(sys.argv[1:])